requirements.txt
langchain
pypdf2
python-dotenv
streamlit
unstructured[all-docs]
pillow
lxml
tiktoken
langchain
langchain-community
langchain-openai
langchain-groq
pymupdf
opencv-python
reportlab
python-pptx
rank-bm25
pysqlite3-binary
chromadb
azure-storage-blob>=12.20.0
azure-identity>=1.17.0
azure-search-documents==11.6.0b12